% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{avgSignals}
\alias{avgSignals}
\title{Takes the average signal of a list of n-dimensional signals.}
\usage{
avgSignals(signal_list, index = TRUE)
}
\arguments{
\item{signal_list}{A list of dataframes with n columns that represent the signals to average.}

\item{index}{Boolean which is true if there is an index column in the dataframes (index column must be the first column).}
}
\value{
A single dataframe that represents the average signal.
}
\description{
Takes the average signal of a list of n-dimensional signals.
}
\examples{
sig1 <- data.frame(x = c(1,2,3,4,5,6,7,8,9,10),
                   y = c(2,3,4,5,4,5,6,7,8,8),
                   z = c(1,1,2,3,2,3,4,3,4,5))
sig2 <- data.frame(x = c(1,2,3,4,5,6,7,8,9,10)+1,
                   y = c(2,3,4,5,4,5,6,7,8,8)+1,
                   z = c(1,1,2,3,2,3,4,3,4,5)+1)
sig3 <- data.frame(x = c(1,2,3,4,5,6,7,8,9,10)+2,
                   y = c(2,3,4,5,4,5,6,7,8,8)+3,
                   z = c(1,1,2,3,2,3,4,3,4,5)+2)
sigList <- list(sig1, sig2, sig3)
avgSignals(sigList, index=FALSE)
}
