% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classifySignals}
\alias{classifySignals}
\title{Classify a signal.}
\usage{
classifySignals(signalOptions, signal, index = TRUE)
}
\arguments{
\item{signalOptions}{A list of dataframes with n columns which represent the options to classify signal as.}

\item{signal}{The signal to classify.}

\item{index}{Boolean which is true if there is an index column in the dataframes (index column must be the first column).}
}
\value{
The index of signalOptions that signal is classified as.
}
\description{
Classify a signal.
}
\examples{
sig1 <- data.frame(x = c(1,2,3,4,5,6,7,8,9,10),
                   y = c(2,3,4,5,4,5,6,7,8,8),
                   z = c(1,1,2,3,2,3,4,3,4,5))
sig2 <- data.frame(x = c(1,2,3,4,5,6,7,8,9,10)+1,
                   y = c(2,3,4,5,4,5,6,7,8,8)+1,
                   z = c(1,1,2,3,2,3,4,3,4,5)+1)
sig3 <- data.frame(x = c(1,2,3,4,5,6,7,8,9,10)+2,
                   y = c(2,3,4,5,4,5,6,7,8,8)+3,
                   z = c(1,1,2,3,2,3,4,3,4,5)+2)
sigList <- list(sig1, sig2, sig3)
classifySignals(sigList, sig1, index=FALSE)
}
