% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classifySignals_highVar}
\alias{classifySignals_highVar}
\title{Classify a signal based on the nSigs highest variance signals.}
\usage{
classifySignals_highVar(signalOptions, signal, nSigs, index = T)
}
\arguments{
\item{signalOptions}{A list of dataframes with n columns which represent the options to classify signal as.}

\item{signal}{The signal to classify.}

\item{nSigs}{The number of highest variance signals to compare by.}

\item{index}{Boolean which is true if there is an index column in the dataframes (index column must be the first column).}
}
\value{
The index of signalOptions that signal is classified as.
}
\description{
Classify a signal based on the nSigs highest variance signals.
}
\examples{
sig1 <- data.frame(x = c(1,2,3,4,5,6,7,8,9,10),
                   y = c(2,3,4,5,4,5,6,7,8,8),
                   z = c(1,1,2,3,2,3,4,3,4,5))
sig2 <- data.frame(x = c(1,2,3,4,5,6,7,8,9,10)+1,
                   y = c(2,3,4,5,4,5,6,7,8,8)+1,
                   z = c(1,1,2,3,2,3,4,3,4,5)+1)
sig3 <- data.frame(x = c(1,2,3,4,5,6,7,8,9,10)+2,
                   y = c(2,3,4,5,4,5,6,7,8,8)+3,
                   z = c(1,1,2,3,2,3,4,3,4,5)+2)
sigList <- list(sig1, sig2, sig3)
classifySignals_highVar(signalOption, sig1, nSigs=2, index=F)
}
