% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compareSignals}
\alias{compareSignals}
\title{Compare two n-dimesional signals}
\usage{
compareSignals(signal1, signal2, index = T)
}
\arguments{
\item{signal1}{A dataframe with n columns.}

\item{signal2}{A dataframe with n columns.}

\item{index}{Boolean which is true if there is an index column in the dataframes (index column must be the first column).}
}
\value{
A numeric representing the distance measure between the two signals.
}
\description{
Compare two n-dimesional signals
}
\examples{
sig1 <- data.frame(x = c(1,2,3,4,5,6,7,8,9,10), y = c(2,3,4,5,4,5,6,7,8,8), z = c(1,1,2,3,2,3,4,3,4,5))
sig2 <- data.frame(x = c(1,2,3,4,5,6,7,8,9,10)+1, y = c(2,3,4,5,4,5,6,7,8,8)+1, z = c(1,1,2,3,2,3,4,3,4,5)+1)
compareSignals(sig1, sig2, index=F)
}
