% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compareSignals_highVar}
\alias{compareSignals_highVar}
\title{Compares the m highest variance signals from two n-dimesional signals (m < n).}
\usage{
compareSignals_highVar(signal1, signal2, nSigs, index = T)
}
\arguments{
\item{signal1}{A dataframe with n columns.}

\item{signal2}{A dataframe with n columns.}

\item{nSigs}{Number of signals to compare. Will choose the nSigs highest variance signals from signal 2 and compare them to the corresponding signals from signal1.}

\item{index}{Boolean which is true if there is an index column in the dataframes (index column must be the first column).}
}
\value{
A numeric representing the distance measure between the two signals.
}
\description{
Compares the m highest variance signals from two n-dimesional signals (m < n).
}
\examples{
sig1 <- data.frame(x = c(1,2,3,4,5,6,7,8,9,10),
                   y = c(2,3,4,5,4,5,6,7,8,8),
                   z = c(1,1,2,3,2,3,4,3,4,5))
sig2 <- data.frame(x = c(1,2,3,4,5,6,7,8,9,10)+1,
                   y = c(2,3,4,5,4,5,6,7,8,8)+1,
                   z = c(1,1,2,3,2,3,4,3,4,5)+1)
compareSignals(sig1, sig2, nSigs=2, index=F)
}
