% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{movingAverage}
\alias{movingAverage}
\title{Smooth a signal with a moving average function}
\usage{
movingAverage(signal, window, index = TRUE)
}
\arguments{
\item{signal}{A dataframe with n columns.}

\item{window}{The size of the moving average window. Each point i has a window of (i-window, i+window).}

\item{index}{Boolean which is true if there is an index column in the dataframes (index column must be the first column).}
}
\value{
A dataframe with n columns representing the smoothed signal.
}
\description{
Smooth a signal with a moving average function
}
\examples{
sig1 <- data.frame(x = c(1,2,3,4,5,6,7,8,9,10),
                   y = c(2,3,4,5,4,5,6,7,8,8),
                   z = c(1,1,2,3,2,3,4,3,4,5))
movingAverage(sig1, 2, index=FALSE)
}
